change jdbc type of crg question and answer point to double

+++ kb-api/src/main/java/com/sb/solutions/api/crg/entity/CrgAnswer.java
@@ -23,7 +23,7 @@ public class CrgAnswer extends BaseEntity<Long> {
     @NotNull(message = "Description is required.")
     private String description;
 
-    private long points;
+    private double points;
 
     private Status status;
